+++ lib/inputs/intents/apply-checklist.intent.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+class ApplyChecklistIntent extends Intent {
+  const ApplyChecklistIntent();
+}